creating encoder and decoder in self supervised contrastive
