+++ src/data_generator.py
@@ -196,82 +196,23 @@ class AutoDataset:
                 output[:,o_idx] = cat_data[:,c_idx]
         return output
 
-class EmbeddingGenerator(torch.nn.Module):
-    """
-        Classical embeddings generator
-        adopted from https://github.com/dreamquark-ai/tabnet/
-    """
 
-    def __init__(self, input_dim, cat_dims, cat_idxs, cat_emb_dim=[]):
-        """ This is an embedding module for an enite set of features
-        Parameters
-        ----------
-        input_dim : int
-            Number of features coming as input (number of columns)
-        cat_dims : list of int
-            Number of modalities for each categorial features
-            If the list is empty, no embeddings will be done
-        cat_idxs : list of int
-            Positional index for each categorical features in inputs
-        cat_emb_dim : int or list of int
-            Embedding dimension for each categorical features
-            If int, the same embdeding dimension will be used for all categorical features
-        """
-        super(EmbeddingGenerator, self).__init__()
-        if cat_dims == [] or cat_idxs == []:
-            self.skip_embedding = True
-            self.post_embed_dim = input_dim
-            return
+def get_cat_info(data):
 
-        # heuristic
-        if (len(cat_emb_dim) == 0):
-            # use heuristic
-            cat_emb_dim = [min(600, round(1.6 * n_cats ** .56)) for n_cats in cat_dims]
+    cat_idxs = []
+    cat_dims = []
 
-        self.skip_embedding = False
-        if isinstance(cat_emb_dim, int):
-            self.cat_emb_dims = [cat_emb_dim]*len(cat_idxs)
-        else:
-            self.cat_emb_dims = cat_emb_dim
+    # get the first row
+    line = data[0]
+    # go through the elements to look for non int/floats
+    for idx in range(len(line)):
+        if isinstance(line[idx], str):
+            cat_idxs.append(idx)
 
-        # check that all embeddings are provided
-        if len(self.cat_emb_dims) != len(cat_dims):
-            msg = """ cat_emb_dim and cat_dims must be lists of same length, got {len(self.cat_emb_dims)}
-                      and {len(cat_dims)}"""
-            raise ValueError(msg)
-        self.post_embed_dim = int(input_dim + np.sum(self.cat_emb_dims) - len(self.cat_emb_dims))
+    # transpose to fo through the rows and get
 
-        self.embeddings = torch.nn.ModuleList()
+    data = data.T
+    for idx in cat_idxs:
+        cat_dims.append(len(np.unique(data[idx])))
 
-        # Sort dims by cat_idx
-        sorted_idxs = np.argsort(cat_idxs)
-        cat_dims = [cat_dims[i] for i in sorted_idxs]
-        self.cat_emb_dims = [self.cat_emb_dims[i] for i in sorted_idxs]
-
-        for cat_dim, emb_dim in zip(cat_dims, self.cat_emb_dims):
-            self.embeddings.append(torch.nn.Embedding(cat_dim, int(emb_dim)))
-        # record continuous indices
-        self.continuous_idx = torch.ones(input_dim, dtype=torch.bool)
-        self.continuous_idx[cat_idxs] = 0
-
-    def forward(self, x):
-        """
-        Apply embdeddings to inputs
-        Inputs should be (batch_size, input_dim)
-        Outputs will be of size (batch_size, self.post_embed_dim)
-        """
-        if self.skip_embedding:
-            # no embeddings required
-            return x
-        cols = []
-        cat_feat_counter = 0
-        for feat_init_idx, is_continuous in enumerate(self.continuous_idx):
-            # Enumerate through continuous idx boolean mask to apply embeddings
-            if is_continuous:
-                cols.append(x[:, feat_init_idx].float().view(-1, 1))
-            else:
-                cols.append(self.embeddings[cat_feat_counter](x[:, feat_init_idx].long()))
-                cat_feat_counter += 1
-        # concat
-        post_embeddings = torch.cat(cols, dim=1)
-        return post_embeddings
+    return cat_idxs, cat_dims
