--- src/test.py
@@ -1,5 +1,0 @@
-from modules import AE
-
-model = AE(50, [50,25],[25,50])
-
-print(model)
